+++ .storybook/main.js
@@ -71,6 +71,8 @@ module.exports = {
           '@blaze-react/checkboxes': path.resolve(__dirname, '../packages/Checkboxes/src'),
           '@blaze-react/chips': path.resolve(__dirname, '../packages/Chips/src'),
           '@blaze-react/themes': path.resolve(__dirname, '../packages/themes/src'),
+          '@blaze-react/modal': path.resolve(__dirname, '../packages/Modal/src'),
+          '@blaze-react/icon': path.resolve(__dirname, '../packages/Icon/src'),
         },
       },
     };