+++ packages/Button/stories/TextButton.stories.tsx
@@ -1,79 +1,57 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React, { useState } from 'react';
-import { BlazeButton } from '../src/BlazeButton';
+import Button from '../src';
 import { adminTheme, frontendTheme, ThemeContext, applyTheme } from '@blaze-react/themes';
+import CustomIcon from '@blaze-react/icon';
+import { args, argTypes } from './shared';
 
 export default {
-  title: '@blaze-react/Button/Text',
-  component: BlazeButton,
-  args: {
-    selectedTheme: 'admin',
-    icon: 'left',
-  },
-  argTypes: {
-    selectedTheme: {
-      options: ['admin', 'frontend'],
-      control: { type: 'radio' },
-      defaultValue: { summary: 'admin' },
-    },
-    theme: {
-      control: false,
-      table: {
-        disable: true,
-      },
-    },
-    icon: {
-      options: ['left', 'right', 'icon-only'],
-      control: { type: 'radio' },
-      defaultValue: { summary: 'left' },
-    },
-  },
+  title: '@blaze-react/Button/All Stories/Text',
+  component: Button,
+  args: args,
+  argTypes: argTypes,
 } as Meta;
 
 const Template: Story = (args) => {
   const selectedTheme = args.selectedTheme === 'admin' ? adminTheme : frontendTheme;
-  const [theme, setTheme] = useState(selectedTheme);
 
   applyTheme(selectedTheme);
 
   return (
-    <ThemeContext.Provider value={{ theme, setTheme }}>
-      <BlazeButton {...args}>
-        <div className="flex items-center justify-center">
-          {args.icon === 'left' ? (
-            <>
-              <svg xmlns="http://www.w3.org/2000/svg" className="w-5 h-5" viewBox="0 0 20 20" fill="currentColor">
-                <path
-                  fillRule="evenodd"
-                  d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z"
-                  clipRule="evenodd"
-                />
-              </svg>
-              {args.icon !== 'icon-only' && <span className="mx-1">Save</span>}
-            </>
-          ) : (
-            <>
-              {args.icon !== 'icon-only' && <span className="mx-1">Save</span>}
+    <Button {...args}>
+      <div className="flex items-center justify-center">
+        {args.icon === 'left' ? (
+          <>
+            <CustomIcon utilities="w-6 h-6 stroke-current fill-current" content={args.svgIcon} />
+            {args.icon !== 'icon-only' && <span className="mx-1">{args.label ?? 'Blaze'}</span>}
+            {args.icon === 'icon-only' && args.icon === 'with-text' && (
+              <span className="mx-1">{args.label ?? 'Blaze'}</span>
+            )}
+          </>
+        ) : (
+          <>
+            {args.icon !== 'icon-only' && <span className="mx-1">{args.label ?? 'Blaze'}</span>}
 
-              <svg xmlns="http://www.w3.org/2000/svg" className="w-5 h-5" viewBox="0 0 20 20" fill="currentColor">
-                <path
-                  fillRule="evenodd"
-                  d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
-                  clipRule="evenodd"
-                />
-              </svg>
-            </>
-          )}
-        </div>
-      </BlazeButton>
-    </ThemeContext.Provider>
+            {args.icon !== 'no-icon' && (
+              <CustomIcon utilities="w-6 h-6 stroke-current fill-current" content={args.svgIcon} />
+            )}
+
+            {args.icon === 'icon-only' && args.icon === 'with-text' && (
+              <span className="mx-1">{args.label ?? 'Blaze'}</span>
+            )}
+          </>
+        )}
+      </div>
+    </Button>
   );
 };
 
 export const Left = Template.bind({});
 export const Right = Template.bind({});
+export const NoIcon = Template.bind({});
 
 Left.args = {
+  icon: 'left',
   theme: adminTheme,
   utilities: `
     font-manrope
@@ -81,13 +59,13 @@ Left.args = {
     text-bold
     rounded-button
     text-button-outlined 
-    w-30
+    min-w-min
     dark:bg-gray-800
     dark:text-button-primary
     hover:text-button-muted
     focus:outline-none
     active:text-button-inverted
-    text-sm`,
+    text-md`,
 };
 
 Right.args = {
@@ -99,11 +77,29 @@ Right.args = {
   text-bold
   rounded-button
   text-button-outlined 
-  w-30
+  min-w-min
   dark:bg-gray-800
   dark:text-button-primary
   hover:text-button-muted
   focus:outline-none
   active:text-button-inverted
-  text-sm`,
+  text-md`,
+};
+
+NoIcon.args = {
+  icon: 'no-icon',
+  theme: adminTheme,
+  utilities: `
+  font-manrope
+  font-medium
+  text-bold
+  rounded-button
+  text-button-outlined 
+  min-w-min
+  dark:bg-gray-800
+  dark:text-button-primary
+  hover:text-button-muted
+  focus:outline-none
+  active:text-button-inverted
+  text-md`,
 };