+++ packages/Button/CHANGELOG.md
@@ -3,6 +3,139 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0-alpha.20](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.19...v0.8.0-alpha.20) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.19](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.18...v0.8.0-alpha.19) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.18](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.17...v0.8.0-alpha.18) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.17](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.16...v0.8.0-alpha.17) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.16](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.15...v0.8.0-alpha.16) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.15](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.14...v0.8.0-alpha.15) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.14](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.13...v0.8.0-alpha.14) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.13](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.12...v0.8.0-alpha.13) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.12](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.11...v0.8.0-alpha.12) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.11](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.10...v0.8.0-alpha.11) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.10](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.9...v0.8.0-alpha.10) (2021-07-13)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.9](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.8...v0.8.0-alpha.9) (2021-07-12)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.8](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.7...v0.8.0-alpha.8) (2021-07-12)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.7](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.6...v0.8.0-alpha.7) (2021-07-12)
+
+**Note:** Version bump only for package @blaze-react/button
+
+
+
+
+
+# [0.8.0-alpha.6](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.3...v0.8.0-alpha.6) (2021-07-09)
+
+
+### Features
+
+* add stories for filled, affix, outlined and text buttons ([9f5cff4](https://github.com/thebyte9/blaze-components-react/commit/9f5cff45b8c6a94ebb5dd7369624bb1dbb892c10))
+* add Tab stories ([13adc96](https://github.com/thebyte9/blaze-components-react/commit/13adc969d8e58179e2ba661039b2e2c4a491e9fd))
+* add variations page ([29abb4f](https://github.com/thebyte9/blaze-components-react/commit/29abb4f048040b2e2ad029a28a6029630692819d))
+
+
+
+
+
 # [0.8.0-alpha.3](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.2...v0.8.0-alpha.3) (2021-06-17)
 
 **Note:** Version bump only for package @blaze-react/button